docs: generate todos without writing a file
